requirements.txt
pandas==1.5.3
numpy==1.21.2
pytest==7.4.3
pytest-cov==4.1.0
black==23.10.1
flake8==6.1.0
openai==1.6.0
python-dotenv==1.0.0
